pyproject.toml
[project]
name = "news-comparison-streamlit"
version = "0.1.0"
description = "Interactive tool to compare Simfoni vs Perplexity news"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27.0",
    "streamlit>=1.54.0",
    "exa-py>=1.0.0",
    "linkup-sdk>=0.2.0",
]
